feat: 로비 씬 꾸미기

+++ Assets/Scripts/Network/PacketUtils.cs
@@ -14,20 +14,19 @@ public class PacketUtils
 {
     public static byte[] SerializePacket<T>(T pPkt, ePacketID pMessageID, int pSequence) where T : IMessage
     {
-        int size = pPkt.CalculateSize() + Marshal.SizeOf(typeof(PacketHeader));
+        Int16 size = (Int16)(pPkt.CalculateSize() + Marshal.SizeOf(typeof(PacketHeader)));
         PacketHeader header = new PacketHeader();
-        header.size = (ushort)size;
+        header.size = size;
         header.id = pMessageID;
         header.sequence = pSequence;
 
-        //Debug.Log("크기: "+ pPkt.CalculateSize()+" : "+ Marshal.SizeOf(typeof(PacketHeader)));
 
         using (MemoryStream ret = new MemoryStream())
         {
             using (BinaryWriter writer = new BinaryWriter(ret, Encoding.Default, true))
             {
                 //패킷 헤더 직렬화
-                writer.Write(header.size);
+                writer.Write((UInt16)header.size);
                 writer.Write((UInt16)header.id);
                 writer.Write(header.sequence);
             }
